+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## [v1.0.3] - unreleased
+
+### Added
+
+- Added global hotkey (press alt-z to activate app and use arrow keys to navigate items)
+- Displays a link to the website in settings when new version is available
 
 ## [v1.0.2] - 2025-11-10